pyproject.toml
[project]
name = "3dtiles-dl-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "python-dotenv>=1.0.0",
    "trimesh>=4.11.0",
]

[dependency-groups]
dev = [
    "pycollada>=0.9.3",
    "pytest>=9.0.2",
]
